add hitbox_handler
