+++ src/handler/enemy_handler.py
@@ -1,21 +1,22 @@
 import random
+
 from objects.enemy import Enemy
 from objects.hitbox import Hitbox
+from handler.hitbox_handler import HitboxHandler
 
-HITBOX_ENABLED = False
+HITBOX_ENABLED = True
 class EnemyHandler():
     def __init__(self, screen_width, screen_height, background):
         self.screen_width = screen_width
         self.screen_height = screen_height
         self.background = background
         self.enemies = []
-        self.hitboxes = []
-        self.inner_hitboxes = []
+
         self.mouse_x = 0
         self.mouse_y = 0
         self.score = 0
+        self.hitbox_handler = HitboxHandler(self.background)
         self.generate_enemies() # spawn enemies
-
 
     def handle_enemies(self, mouse_x, mouse_y, hud, dt):
         self.mouse_x = mouse_x
@@ -23,17 +24,9 @@ class EnemyHandler():
         self.hud = hud
         for index, enemy in enumerate(self.enemies):
             if HITBOX_ENABLED:
-                self.handle_hitboxes(enemy, index)
+                self.hitbox_handler.handle_hitboxes(enemy, index)
             self.adjust_enemy_position(enemy, index, dt)
             self.check_for_collisions(enemy, index)
-
-    
-    def handle_hitboxes(self, enemy, index):
-        hitbox = self.hitboxes[index]
-        inner_hitbox = self.inner_hitboxes[index]
-        hitbox.adjust_hitbox_position(enemy, hitbox, index, -2)
-        inner_hitbox.adjust_hitbox_position(enemy, inner_hitbox, index, 0)
-
 
     def generate_enemies(self, amount=100):
         dummy_enemy = Enemy(0, 100, 1, 1, self.background)
@@ -43,14 +36,8 @@ class EnemyHandler():
             y = self.screen_height
             newEnemy = Enemy(enemySpeed, 100, x, y, self.background)
             if HITBOX_ENABLED:
-                self.generate_hitbox_for_enemy(newEnemy)
+                self.hitbox_handler.generate_hitbox_for_enemy(newEnemy)
             self.enemies.append(newEnemy)
-
-    
-    def generate_hitbox_for_enemy(self, newEnemy):
-        self.hitboxes.append(Hitbox(newEnemy.speed, newEnemy.image_width + 4, newEnemy.image_height + 4, newEnemy.x_pos - 5, newEnemy.y_pos - 5, self.background, (255, 100, 120)))
-        self.inner_hitboxes.append(Hitbox(newEnemy.speed, newEnemy.image_width, newEnemy.image_height, newEnemy.x_pos, newEnemy.y_pos, self.background, (230, 230, 230)))
-
 
     def adjust_enemy_position(self, enemy, index, dt):
         if enemy.y_pos <= 0:
@@ -58,7 +45,6 @@ class EnemyHandler():
         else:
             enemy.draw(enemy.x_pos, enemy.y_pos)
             enemy.update(dt)
-
 
     def check_for_collisions(self, enemy, index):
         if (self.mouse_x >= int(enemy.x_pos)) and (self.mouse_x <= int(enemy.x_pos) + (enemy.image_width)):
